+++ release-notes.md
@@ -1,3 +1,13 @@
+## [1.2.5](https://github.com/Vastmindz-Public-Repository/Web-SDK/tree/v1.2.5) (2022-10-25)
+
+> Description
+
+### New Features
+* add config param skipSocketWhenBadFaceConditions
+* add config param skipSocketWhenBadLightConditions
+* make width, height, useFrontCamera of camera instance public
+* make rppgCamera, rppgSocket, rppgTracker instances public
+
 ## [1.2.4](https://github.com/Vastmindz-Public-Repository/Web-SDK/tree/v1.2.4) (2022-10-19)
 
 > Description